--- helpers/schema.js
@@ -1,17 +1,0 @@
-// import Joi from "joi";
-
-// export const schemaAdd = Joi.object({
-//   name: Joi.string().required(),
-//   email: Joi.string().required(),
-//   phone: Joi.string().required(),
-// });
-
-// export const schemaUpdate = Joi.object({
-//   name: Joi.string(),
-//   email: Joi.string(),
-//   phone: Joi.string(),
-// }).min(1);
-
-// export const favoriteSchema = Joi.object({
-//   favorite: Joi.bool(),
-// });
