+++ server.js
@@ -1,9 +1,6 @@
 import { app } from "./app.js";
 import mongoose from "mongoose";
 
-// const DB_HOST =
-//   "mongodb+srv://simakov:hJ56BT8zt7chUBeV@cluster0.zmwfuhu.mongodb.net/Contacts_reader?retryWrites=true&w=majority";
-
 const { DB_HOST, PORT = 3000 } = process.env;
 
 mongoose
@@ -13,7 +10,3 @@ mongoose
     console.log(error.message);
     process.exit(1);
   });
-
-// app.listen(3000, () => {
-//   console.log("Server running. Use our API on port: 3000");
-// });
